feat(panel): Agents hub — Fleet and Conversations tabs, /a2a redirect, DM quick-action
