+++ panel/src/components/a2a/a2a-new-dm-dialog.tsx
@@ -39,6 +39,13 @@ interface A2ANewDmDialogProps {
   /** Called with the new (or reopened) conversation's id once the CEO's
    * first message is sent — the caller selects/opens it in the page. */
   onCreated: (conversationId: string) => void;
+  /** Controlled-open pair — omit both for the default uncontrolled trigger-
+   * button behavior (internal state). Pass both to drive the dialog from
+   * outside, e.g. the agent card's DM quick-action deep link. */
+  open?: boolean;
+  onOpenChange?: (open: boolean) => void;
+  /** Preselects this agent as the target whenever the dialog opens. */
+  initialTarget?: string | null;
 }
 
 /**
@@ -46,8 +53,15 @@ interface A2ANewDmDialogProps {
  * classic list only ever show conversations that already exist; this is the
  * one surface that creates one, addressed to any agent (never itself).
  */
-export function A2ANewDmDialog({ onCreated }: A2ANewDmDialogProps) {
-  const [open, setOpen] = useState(false);
+export function A2ANewDmDialog({
+  onCreated,
+  open: openProp,
+  onOpenChange,
+  initialTarget,
+}: A2ANewDmDialogProps) {
+  const [internalOpen, setInternalOpen] = useState(false);
+  const open = openProp ?? internalOpen;
+  const setOpen = onOpenChange ?? setInternalOpen;
   const [targetAgent, setTargetAgent] = useState<string | null>(null);
   const [message, setMessage] = useState("");
   const create = useCreateCeoConversation();
@@ -56,6 +70,14 @@ export function A2ANewDmDialog({ onCreated }: A2ANewDmDialogProps) {
     setTargetAgent(null);
     setMessage("");
   };
+
+  // Preselect on the open transition (render-phase adjustment, not an
+  // effect — same idiom as the connection-banner reset in a2a-view.tsx).
+  const [wasOpen, setWasOpen] = useState(open);
+  if (open !== wasOpen) {
+    setWasOpen(open);
+    if (open && initialTarget) setTargetAgent(initialTarget);
+  }
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
